+++ src/template/Template.ts
@@ -137,9 +137,9 @@ export class Template implements ITemplate {
             new TemplateVariable({
                 name: BuiltinVariables.CWD,
                 defaultValue: scopeCwd,
-                index: 1000,
                 ui: {
                     message: `Output directory`,
+                    index: 1000,
                 },
             }),
         ]);